refac: style 추출
