+++ test-web/src/components/DataListView/style/ordertableheadstyle.js
@@ -1,0 +1,18 @@
+const style = {
+    tableDelHeader: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '6px',
+      fontWeight: '600',
+      color: 'black',
+    },
+    tableHeader: {
+      display: 'flex',
+      alignItems: 'center',
+      padding: '6px',
+      fontWeight: '600',
+      color: '#90a4ae',
+    },
+  };
+
+export default style;
